+++ web/src/features/map/map-utils/getApproximateFeature.ts
